+++ SuperMarioBros-3/FirePiranha.h
@@ -30,7 +30,7 @@ public:
 	bool vanish;
 
 	CFirePiranha(CMario* mario, int piranhaType);
-	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 	virtual void SetState(int state);
