+++ SuperMarioBros-3/Fireball.h
@@ -5,7 +5,7 @@ class CFireball : public CGameObject
 {
 public:
 	CFireball(D3DXVECTOR2 position, int nx);
-	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 	virtual void Render();
 	~CFireball();
