+++ SuperMarioBros-3/GreenPiranha.cpp
@@ -8,11 +8,11 @@ CGreenPiranha::CGreenPiranha(CMario* mario)
 	SetState(FIRE_PIRANHA_STATE_MOVE_UP);
 }
 
-void CGreenPiranha::Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects)
+void CGreenPiranha::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 {
 	CGameObject::Update(dt);
 	y += dy;
-	
+
 	if (effect)
 		effect->Update(dt, coObjects);
