+++ SuperMarioBros-3/IceFlower.h
@@ -4,11 +4,11 @@
 class CIceFlower : public CGameObject
 {
 public:
-	ULONGLONG existingTime;
+	DWORD existingTime;
 	float minPosY;
 
 	CIceFlower(float brickX, float brickY);
-	virtual void Update(ULONGLONG dt, vector<LPGAMEOBJECT>* coObjects);
+	virtual void Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects);
 	virtual void Render();
 	virtual void GetBoundingBox(float& l, float& t, float& r, float& b);
 };
