+++ SuperMarioBros-3/main.cpp
@@ -47,7 +47,7 @@ LRESULT CALLBACK WinProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
 	Update world status for this frame
 	dt: time period between beginning of last frame and beginning of this frame
 */
-void Update(ULONGLONG dt)
+void Update(DWORD dt)
 {
 	CGame::GetInstance()->GetCurrentScene()->Update(dt);
 }
@@ -115,7 +115,7 @@ HWND CreateGameWindow(HINSTANCE hInstance, int nCmdShow, int ScreenWidth, int Sc
 	if (!hWnd) 
 	{
 		OutputDebugString(L"[ERROR] CreateWindow failed");
-		ULONGLONG ErrCode = GetLastError();
+		DWORD ErrCode = GetLastError();
 		return FALSE;
 	}
 
@@ -129,8 +129,8 @@ int Run()
 {
 	MSG msg;
 	int done = 0;
-	ULONGLONG frameStart = GetTickCount64();
-	ULONGLONG tickPerFrame = 1000 / MAX_FRAME_RATE;
+	DWORD frameStart = GetTickCount64();
+	DWORD tickPerFrame = 1000 / MAX_FRAME_RATE;
 
 	while (!done)
 	{
@@ -142,11 +142,11 @@ int Run()
 			DispatchMessage(&msg);
 		}
 
-		ULONGLONG now = GetTickCount64();
+		DWORD now = GetTickCount64();
 
 		// dt: the time between (beginning of last frame) and now
 		// this frame: the frame we are about to render
-		ULONGLONG dt = now - frameStart;
+		DWORD dt = now - frameStart;
 
 		if (dt >= tickPerFrame)
 		{
